backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
bcrypt==4.0.1
Pillow==10.1.0
google-cloud-storage==2.18.2
cloud-sql-python-connector[pg8000]==1.13.0

# AI Dependencies
torch>=2.2.0 --index-url https://download.pytorch.org/whl/cu121
diffusers>=0.26.0
transformers>=4.38.0
accelerate>=0.27.0
opencv-python>=4.9.0
rembg>=2.0.50
onnxruntime>=1.17.0
scikit-image>=0.22.0
numpy<2.0.0
